moved the traver and parsing functionality to traverseParseFunctions file
